Rename and move phases to qualifier
